requirements.txt
python-telegram-bot
transformers
pinecone-client
sentence-transformers
torch
